test(cli): move model family httpapi coverage
